proper erros messages, finish searchpage merge
