+++ src/hooks/use-fetch.jsx
@@ -4,7 +4,7 @@ import data from '../../data/db.json';
 const useFetchFile = () => {
     const [services, setServices] = useState(null);
     const [businesses, setBusinesses] = useState(null);
-    const [errors, setErrors] = useState(null);
+    const [errors, setErrors] = useState([]);
     const [isLoading, setIsLoading] = useState(true);
 
     useEffect(() => {
@@ -13,13 +13,13 @@ const useFetchFile = () => {
                 const newErrors = []; // Local array to collect any errors
 
                 if (!data.services) {
-                    newErrors.push("Services not found");
+                    newErrors.push({message: "Services not found"});
                 } else {
                     setServices(data.services);
                 }
 
                 if (!data.businesses) {
-                    newErrors.push("Businesses not found");
+                    newErrors.push({message: "Businesses not found"});
                 } else {
                     setBusinesses(data.businesses);
                 }
